+++ packages/liferay-theme-tasks/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Next Release
+* No changes
+***
+
+## v1.1.0 - April 27, 2016
 * New: add deploy:gogo task for faster full deploy for 7.0 themes
 * Fix: fixed issue where theme names were being truncated on deploy
-***
 
 ## v1.0.9 - March 30, 2016
 * New: allow devs to pass configuration options to sass compiler